requirements.txt
python-dotenv
streamlit
pyswisseph
pytz
pandas
numpy
python-dateutil
geopy
jinja2
openai 
kerykeion
pyjhora
svgwrite
langcodes